requirements.txt
astroid==2.9.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
pycodestyle==2.8.0
pylint==2.12.2
python-decouple==3.5
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
wrapt==1.13.3
